arreglo backups

+++ DAM_PROYECTO_PABLO_CARBONERO/app/src/main/java/com/example/dam_proyecto_pablo_carbonero/lib/domain/usecases/firebaseUseCases/DownloadBackupUseCase.kt
@@ -4,6 +4,7 @@ import com.example.dam_proyecto_pablo_carbonero.lib.data.local.dao.MusicNoteDao
 import com.example.dam_proyecto_pablo_carbonero.lib.data.local.dao.SongDao
 import com.example.dam_proyecto_pablo_carbonero.lib.data.local.dao.TuningDao
 import com.example.dam_proyecto_pablo_carbonero.lib.data.local.dao.TuningMusicNoteDao
+import com.example.dam_proyecto_pablo_carbonero.lib.domain.repositories.FirebaseRepository
 import com.example.dam_proyecto_pablo_carbonero.lib.domain.repositories.MusicNoteRepository
 import com.example.dam_proyecto_pablo_carbonero.lib.domain.repositories.SongRepository
 import com.example.dam_proyecto_pablo_carbonero.lib.domain.repositories.TuningMusicNoteRepository
@@ -30,49 +31,47 @@ class DownloadBackupUseCase @Inject constructor(
     private val tuningRepository: TuningRepository,
     private val tuningMusicNoteRepository: TuningMusicNoteRepository,
     private val songRepository: SongRepository,
-    private val prefsRepository: UserPreferencesRepository
+    private val prefsRepository: UserPreferencesRepository,
+    private val firebaseRepository: FirebaseRepository
 ) : UseCase<Unit, Boolean> {
     override suspend fun call(param: Unit): Boolean {
-        val firestore = Firebase.firestore
-        val userId = Firebase.auth.currentUser?.uid ?: return false
 
-        firestore.collection("backups")
-            .document(userId)
-            .get()
-            .addOnSuccessListener { doc ->
-                if (doc.exists()) {
-                    val notes = doc["MusicNote"] as? List<Map<String, Any>>
-                    val songs = doc["Song"] as? List<Map<String, Any>>
-                    val tunings = doc["Tuning"] as? List<Map<String, Any>>
-                    val tuningMusicNote = doc["TuningMusicNote"] as? List<Map<String, Any>>
+        val doc = firebaseRepository.downloadBackup()
+
+        if (doc != null && doc.exists()) {
+            val notes = doc["MusicNote"] as? List<Map<String, Any>>
+            val songs = doc["Song"] as? List<Map<String, Any>>
+            val tunings = doc["Tuning"] as? List<Map<String, Any>>
+            val tuningMusicNote = doc["TuningMusicNote"] as? List<Map<String, Any>>
 
 
-                    // Convertir a tus clases Kotlin
-                    val restoredNotes = notes?.map { it.toMusicNote() }
-                    val restoredTunings = tunings?.map { it.toTuning() }
-                    val restoredTuningMusicNote = tuningMusicNote?.map { it.toTuningMusicNote() }
-                    val restoredSong = songs?.map { it.toSong() }
+            // Convertir a tus clases Kotlin
+            val restoredNotes = notes?.map { it.toMusicNote() }
+            val restoredTunings = tunings?.map { it.toTuning() }
+            val restoredTuningMusicNote = tuningMusicNote?.map { it.toTuningMusicNote() }
+            val restoredSong = songs?.map { it.toSong() }
 
-                    // Guardar en Room reemplazando lo anterior
-                    CoroutineScope(Dispatchers.IO).launch {
-                        tuningMusicNoteRepository.deleteAllTuningMusicNote()
-                        musicNoteRepository.deleteAllMusicNotes()
-                        tuningRepository.deleteAllTunings()
-                        songRepository.deleteAllSongs()
 
-                        musicNoteRepository.insertAllMusicNotes(restoredNotes ?: emptyList())
-                        tuningRepository.insertAllTuning(restoredTunings ?: emptyList())
-                        tuningMusicNoteRepository.insertAllTuningMusicNote(restoredTuningMusicNote ?: emptyList())
-                        songRepository.insertAllSongs(restoredSong ?: emptyList())
+            tuningMusicNoteRepository.deleteAllTuningMusicNote()
+            musicNoteRepository.deleteAllMusicNotes()
+            tuningRepository.deleteAllTunings()
+            songRepository.deleteAllSongs()
 
-                        val date = Timestamp.now().toDate()
-                        val formatter = SimpleDateFormat("yyyy-MM-dd' 'HH:mm:ss", Locale.getDefault())
-                        prefsRepository.setLastModificationDate(formatter.format(date))
-                    }
-                }
-            }
+            musicNoteRepository.insertAllMusicNotes(restoredNotes ?: emptyList())
+            tuningRepository.insertAllTuning(restoredTunings ?: emptyList())
+            tuningMusicNoteRepository.insertAllTuningMusicNote(
+                restoredTuningMusicNote ?: emptyList()
+            )
+            songRepository.insertAllSongs(restoredSong ?: emptyList())
 
-        return true
+            val date = Timestamp.now().toDate()
+            val formatter = SimpleDateFormat("yyyy-MM-dd' 'HH:mm:ss", Locale.getDefault())
+            prefsRepository.setLastModificationDate(formatter.format(date))
+
+            return true
+        }
+        else {
+            return false
+        }
     }
-
 }
